requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
bidict==0.23.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
debugpy==1.8.11
fastapi==0.115.5
filetype==1.2.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.8
langchain-core==0.3.25
langchain-google-genai==2.0.7
langchain-ollama==0.2.1
langchain-text-splitters==0.3.3
langgraph==0.2.56
langgraph-checkpoint==2.0.9
langgraph-checkpoint-postgres==2.0.9
langgraph-sdk==0.1.47
langsmith==0.1.147
loguru==0.7.3
Mako==1.3.8
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
netifaces==0.10.6
numpy==1.26.4
ollama==0.4.4
orjson==3.10.12
packaging==24.2
pipdeptree==2.24.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-engineio==4.11.1
python-multipart==0.0.19
python-socketio==5.12.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==3.3
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
wsproto==1.2.0
yarl==1.18.3
